add delay in user_ddriver for optimizations test

+++ fs/simplefs/tests/test.sh
@@ -1,5 +1,16 @@
 #!/bin/bash
 
+cd ../../../tests/test_ddriver || exit
+out=$(./compile_and_run.sh)
+if ! echo "$out" | grep "Test Pass :)"; then
+    echo "-- Output:"
+    echo "$out"
+    echo "请先通过实验一 (确保能够理解ddriver)"
+    echo "Score: 0"
+    exit 1
+fi 
+cd - || exit
+
 read -r -p "请输入测试方式[N(基础功能测试) / E(进阶功能测试) / S(分阶段测试)]: " TEST_METHOD
 
 rm mnt -rf
